Refactor into a class pattern
This offers a lot more readability and more separation of concerns. It
also allow more reuse and places the linkages to the DOM declared up
front.

For now this executes on document ready. However, in the future the job
of instantiating should be pulled out of this file and handled in a file
higher up the module chain. That way this class can be tested in
isolation with out inadvertent initialization.